ref: 入室処理を別コンポーネントに移動
