+++ src/app/rooms/[id]/page.tsx
@@ -1,6 +1,7 @@
 import { Board } from './_components/Board';
 import { Hand } from './_components/Hand';
 import { ShareButton } from './_components/ShareButton';
+import { RoomEntry } from './_components/RoomEntry';
 import Link from 'next/link';
 import { Sparkles, Home } from 'lucide-react';
 import { Metadata } from 'next';
@@ -32,6 +33,7 @@ export default async function RoomPage({ params }: Props) {
 
   return (
     <div className="min-h-screen bg-linear-to-br from-indigo-50 via-white to-purple-50">
+      <RoomEntry roomId={roomId} />
       <header className="bg-white/80 backdrop-blur-md shadow-lg sticky top-0 z-20 border-b border-gray-200">
         <div className="max-w-7xl mx-auto px-3 sm:px-4">
           <div className="flex items-center justify-between py-1.5 gap-2">
